% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesTSM_package.R
\docType{package}
\name{BayesTSM-package}
\alias{BayesTSM}
\alias{BayesTSM-package}
\title{BayesTSM: Bayesian Progressive Three State Model With Censoring due to Intervention}
\description{
Estimates a Bayesian progressive Three State Model with censoring due to intervention as described in Klausch et al. (2023) \url{https://doi.org/10.1214/22-AOAS1669}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasklausch2/bayestsm}
  \item Report bugs at \url{https://github.com/thomasklausch2/bayestsm/issues}
}

}
\author{
\strong{Maintainer}: Thomas Klausch \email{t.klausch@amsterdamumc.nl}

Authors:
\itemize{
  \item Thomas Klausch \email{t.klausch@amsterdamumc.nl}
}

}
\keyword{internal}
