% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bayestsm.R
\name{plot.bayestsm}
\alias{plot.bayestsm}
\title{Plot method for bayestsm objects}
\usage{
\method{plot}{bayestsm}(x, warmup = 0, plot.X = TRUE, plot.S = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{"bayestsm"}.}

\item{warmup}{Number of initial MCMC iterations to discard.}

\item{plot.X}{Logical; whether trace plots for the parameters of the x-model should be made.}

\item{plot.S}{Logical; whether trace plots for the parameters of the s-model should be made.}

\item{...}{Additional arguments passed to plotting functions.}
}
\description{
Plot method for bayestsm objects
}
