% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bayestsm.R
\name{summary.bayestsm}
\alias{summary.bayestsm}
\title{Summary method for bayestsm objects}
\usage{
\method{summary}{bayestsm}(object, warmup = 0, probs = c(0.025, 0.5, 0.975), ...)
}
\arguments{
\item{object}{An object of class \code{"bayestsm"}.}

\item{warmup}{Number of initial MCMC iterations to discard.}

\item{probs}{Numeric vector of posterior quantiles.}

\item{...}{Additional arguments, currently unused.}
}
\description{
Summary method for bayestsm objects
}
